Temporary rename file to fix casing issue.
