+++ plugins/woocommerce-admin/client/marketing/hooks/oldUseInstalledPluginsWithoutChannels.ts
